requirements.txt.txt
streamlit
google-generativeai
requests
pypdf